+++ DatadogRUM/Tests/Mocks/RUMFeatureMocks.swift
@@ -776,7 +776,8 @@ extension RUMScopeDependencies {
         viewCache: ViewCache = ViewCache(dateProvider: SystemDateProvider()),
         fatalErrorContext: FatalErrorContextNotifying = FatalErrorContextNotifierMock(),
         sessionEndedMetric: SessionEndedMetricController = SessionEndedMetricController(telemetry: NOPTelemetry(), sampleRate: 0),
-        watchdogTermination: WatchdogTerminationMonitor = .mockRandom()
+        watchdogTermination: WatchdogTerminationMonitor = .mockRandom(),
+        networkSettledMetricFactory: @escaping (Date) -> TTNSMetricTracking = { _ in TTNSMetricMock() }
     ) -> RUMScopeDependencies {
         return RUMScopeDependencies(
             featureScope: featureScope,
@@ -795,7 +796,8 @@ extension RUMScopeDependencies {
             viewCache: viewCache,
             fatalErrorContext: fatalErrorContext,
             sessionEndedMetric: sessionEndedMetric,
-            watchdogTermination: watchdogTermination
+            watchdogTermination: watchdogTermination,
+            networkSettledMetricFactory: networkSettledMetricFactory
         )
     }
 
@@ -816,7 +818,8 @@ extension RUMScopeDependencies {
         viewCache: ViewCache? = nil,
         fatalErrorContext: FatalErrorContextNotifying? = nil,
         sessionEndedMetric: SessionEndedMetricController? = nil,
-        watchdogTermination: WatchdogTerminationMonitor? = nil
+        watchdogTermination: WatchdogTerminationMonitor? = nil,
+        networkSettledMetricFactory: ((Date) -> TTNSMetricTracking)? = nil
     ) -> RUMScopeDependencies {
         return RUMScopeDependencies(
             featureScope: self.featureScope,
@@ -835,7 +838,8 @@ extension RUMScopeDependencies {
             viewCache: viewCache ?? self.viewCache,
             fatalErrorContext: fatalErrorContext ?? self.fatalErrorContext,
             sessionEndedMetric: sessionEndedMetric ?? self.sessionEndedMetric,
-            watchdogTermination: watchdogTermination
+            watchdogTermination: watchdogTermination,
+            networkSettledMetricFactory: networkSettledMetricFactory ?? self.networkSettledMetricFactory
         )
     }
 }
@@ -978,6 +982,7 @@ extension RUMResourceScope {
         isFirstPartyResource: Bool? = nil,
         resourceKindBasedOnRequest: RUMResourceType? = nil,
         spanContext: RUMSpanContext? = .mockAny(),
+        networkSettledMetric: TTNSMetricTracking = TTNSMetric(viewStartDate: .mockAny()),
         onResourceEventSent: @escaping () -> Void = {},
         onErrorEventSent: @escaping () -> Void = {}
     ) -> RUMResourceScope {
@@ -992,6 +997,7 @@ extension RUMResourceScope {
             httpMethod: httpMethod,
             resourceKindBasedOnRequest: resourceKindBasedOnRequest,
             spanContext: spanContext,
+            networkSettledMetric: networkSettledMetric,
             onResourceEventSent: onResourceEventSent,
             onErrorEventSent: onErrorEventSent
         )
@@ -1182,5 +1188,38 @@ extension AppHang.BacktraceGenerationResult: AnyMockable, RandomMockable {
             .failed,
             .notAvailable
         ].randomElement()!
+    }
+}
+
+// MARK: - View Loading Metrics
+
+internal class TTNSMetricMock: TTNSMetricTracking {
+    /// Tracks calls to `trackResourceStart(at:resourceID:)`.
+    var resourceStartDates: [RUMUUID: Date] = [:]
+    /// Tracks calls to `trackResourceEnd(at:resourceID:resourceDuration:)`.
+    var resourceEndDates: [RUMUUID: (Date, TimeInterval?)] = [:]
+    /// Tracks if `trackViewWasStopped()` was called.
+    var viewWasStopped = false
+    /// Mocked value returned by this metric.
+    var value: TimeInterval?
+
+    init(value: TimeInterval? = nil) {
+        self.value = value
+    }
+
+    func trackResourceStart(at startDate: Date, resourceID: RUMUUID) {
+        resourceStartDates[resourceID] = startDate
+    }
+
+    func trackResourceEnd(at endDate: Date, resourceID: RUMUUID, resourceDuration: TimeInterval?) {
+        resourceEndDates[resourceID] = (endDate, resourceDuration)
+    }
+
+    func trackViewWasStopped() {
+        viewWasStopped = true
+    }
+
+    func value(at time: Date, appStateHistory: AppStateHistory) -> TimeInterval? {
+        return value
     }
 }